+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - It is now verified on all resources that the UUID really belongs to the respective job type. Such requests will be answered with HTTP code 400 and a respective error message.
 - Detect and prevent the definition of multiple APIs or jobs with the same job type. The second API/job with the same job type will cause an `IllegalArgumentException` to be thrown.
 - The main method `JobFramework` was split into two methods to simplify the API: Instead of taking an `Either<Application, CoroutineScope>`, there is now one method taking an `Application` and one taking a `CoroutineScope`. 
+- Renamed method `fetchStati` of `JobPersistence` to `fetchStates`
 
 
 ## [1.0.0-RC2] - 2023-05-08